+++ fuel-tx/src/builder.rs
@@ -58,17 +58,18 @@ use fuel_types::{
 };
 
 pub trait BuildableAloc
-    where
-        Self: Default + Clone + Executable + Chargeable + field::Policies + Into<Transaction>,
-{}
+where
+    Self: Default + Clone + Executable + Chargeable + field::Policies + Into<Transaction>,
+{
+}
 
 pub trait BuildableStd: Signable + Cacheable {}
 
 pub trait BuildableSet: BuildableAloc + BuildableStd {}
 
 pub trait Buildable
-    where
-        Self: BuildableSet,
+where
+    Self: BuildableSet,
 {
     /// Append an input to the transaction
     fn add_input(&mut self, input: Input) {
@@ -82,8 +83,8 @@ pub trait Buildable
 
     /// Set the `Script`'s gas limit
     fn set_script_gas_limit(&mut self, limit: Word)
-        where
-            Self: field::ScriptGasLimit,
+    where
+        Self: field::ScriptGasLimit,
     {
         *self.script_gas_limit_mut() = limit;
     }
@@ -91,7 +92,8 @@ pub trait Buildable
 
 impl<T> BuildableAloc for T where
     Self: Default + Clone + Executable + Chargeable + field::Policies + Into<Transaction>
-{}
+{
+}
 
 impl<T> BuildableStd for T where T: Signable + Cacheable {}
 
@@ -103,8 +105,6 @@ impl<T> Buildable for T where T: BuildableSet {}
 pub struct TransactionBuilder<Tx> {
     tx: Tx,
 
-    should_prepare_script: bool,
-    should_prepare_predicate: bool,
     params: ConsensusParameters,
 
     // We take the key by reference so this lib won't have the responsibility to properly
@@ -126,12 +126,7 @@ impl TransactionBuilder<Script> {
             receipts_root: Default::default(),
             metadata: None,
         };
-
-        let mut slf = Self::with_tx(tx);
-
-        slf.prepare_script(true);
-
-        slf
+        Self::with_tx(tx)
     }
 }
 
@@ -190,14 +185,10 @@ impl TransactionBuilder<Mint> {
 
 impl<Tx> TransactionBuilder<Tx> {
     fn with_tx(tx: Tx) -> Self {
-        let should_prepare_script = false;
-        let should_prepare_predicate = false;
         let sign_keys = BTreeMap::new();
 
         Self {
             tx,
-            should_prepare_script,
-            should_prepare_predicate,
             params: ConsensusParameters::standard(),
             sign_keys,
         }
@@ -279,17 +270,7 @@ impl<Tx> TransactionBuilder<Tx> {
 }
 
 impl<Tx: Buildable> TransactionBuilder<Tx> {
-    pub fn prepare_script(&mut self, should_prepare_script: bool) -> &mut Self {
-        self.should_prepare_script = should_prepare_script;
-        self
-    }
-
-    pub fn prepare_predicate(&mut self, should_prepare_predicate: bool) -> &mut Self {
-        self.should_prepare_predicate = should_prepare_predicate;
-        self
-    }
-
-    pub fn sign_keys(&self) -> impl Iterator<Item=&SecretKey> {
+    pub fn sign_keys(&self) -> impl Iterator<Item = &SecretKey> {
         self.sign_keys.keys()
     }
 
@@ -299,8 +280,8 @@ impl<Tx: Buildable> TransactionBuilder<Tx> {
     }
 
     pub fn script_gas_limit(&mut self, gas_limit: Word) -> &mut Self
-        where
-            Tx: field::ScriptGasLimit,
+    where
+        Tx: field::ScriptGasLimit,
     {
         self.tx.set_script_gas_limit(gas_limit);
 
@@ -433,19 +414,8 @@ impl<Tx: Buildable> TransactionBuilder<Tx> {
         *witness_index
     }
 
-    fn prepare_finalize(&mut self) {
-        if self.should_prepare_predicate {
-            self.tx.prepare_init_predicate();
-        }
-
-        if self.should_prepare_script {
-            self.tx.prepare_init_script();
-        }
-    }
-
     fn finalize_inner(&self) -> Tx {
         let mut new = self.clone();
-        new.prepare_finalize();
 
         let mut tx = new.tx;
 
@@ -461,7 +431,6 @@ impl<Tx: Buildable> TransactionBuilder<Tx> {
 
     pub fn finalize_without_signature_inner(&self) -> Tx {
         let mut new = self.clone();
-        new.prepare_finalize();
 
         let mut tx = new.tx;
 
@@ -519,9 +488,9 @@ impl Finalizable<Script> for TransactionBuilder<Script> {
 }
 
 impl<Tx> TransactionBuilder<Tx>
-    where
-        Self: Finalizable<Tx>,
-        Transaction: From<Tx>,
+where
+    Self: Finalizable<Tx>,
+    Transaction: From<Tx>,
 {
     pub fn finalize_as_transaction(&mut self) -> Transaction {
         self.finalize().into()