+++ fuel-tx/src/transaction/fee.rs
@@ -1,6 +1,9 @@
 use crate::{
     field,
-    field::WitnessLimit,
+    field::{
+        Tip,
+        WitnessLimit,
+    },
     input::{
         coin::{
             CoinPredicate,
@@ -13,6 +16,7 @@ use crate::{
             MessageDataSigned,
         },
     },
+    policies::PolicyType,
     FeeParameters,
     GasCosts,
     Input,
@@ -151,11 +155,13 @@ pub trait Chargeable: field::Inputs + field::Witnesses + field::Policies {
         fee: &FeeParameters,
         gas_price: Word,
     ) -> u128 {
-        gas_to_fee(
+        let tip = self.tip();
+        let gas_fee = gas_to_fee(
             self.min_gas(gas_costs, fee),
             gas_price,
             fee.gas_price_factor,
-        )
+        );
+        gas_fee.saturating_add(tip as u128)
     }
 
     /// Returns the maximum possible fee after the end of transaction execution.
@@ -167,11 +173,13 @@ pub trait Chargeable: field::Inputs + field::Witnesses + field::Policies {
         fee: &FeeParameters,
         gas_price: Word,
     ) -> u128 {
-        gas_to_fee(
+        let tip = self.tip();
+        let gas_fee = gas_to_fee(
             self.max_gas(gas_costs, fee),
             gas_price,
             fee.gas_price_factor,
-        )
+        );
+        gas_fee.saturating_add(tip as u128)
     }
 
     /// Returns the fee amount that can be refunded back based on the `used_gas` and
@@ -190,7 +198,9 @@ pub trait Chargeable: field::Inputs + field::Witnesses + field::Policies {
         let min_gas = self.min_gas(gas_costs, fee);
 
         let total_used_gas = min_gas.saturating_add(used_gas);
-        let used_fee = gas_to_fee(total_used_gas, gas_price, fee.gas_price_factor);
+        let tip = self.policies().get(PolicyType::Tip).unwrap_or(0);
+        let used_fee = gas_to_fee(total_used_gas, gas_price, fee.gas_price_factor)
+            .saturating_add(tip as u128);
 
         let refund = self
             .max_fee(gas_costs, fee, gas_price)